Added in preconditions of link methods that from and to are not equal. Added new errors for closed graphs and nodes. Added new Error transformer for DAGs and nodes.
Changed that nodes need to specify, in which graphs they are. (var n: Node[g]) means that n is either null or is in an initialized g. If g is a local variable, that has not yet been initialized, n is null. _ means all existing graphs, one can specify multiple graphs (var n: Node[g0, g1]). This specification will always be checked after method calls, and local variable assigns, if they change the node n or a graph, in which n may be.

Added possibility to write Graph(n, m).
Changed SimpleInfo comments, such that the translated code is more readable.
Added new function IS_INITIALIZED(g), which checks if the local variable g is initialized. If g is not a local variable, IS_INITIALIZED is an undefined function.

Fixed several bugs in the reaching definitions analysis.

With the new way of declaring nodes, we can try to optimize the graph, on which the update is applied. (UNLINK(UNIVERSE, n)) could be first translated to (UNLINK(g, n)), and then dependent on the topology of g.

Changed that update invariants are only checked, if there is a field __CONFIG_OUROBOROS_UPDATE_INVARIANTS: Bool.

Added type checking for condExps and EmptySets.